app/src/requirements.txt
cachetools==5.5.0 ; python_version >= "3.12" and python_version < "4.0"
certifi==2024.8.30 ; python_version >= "3.12" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.12" and python_version < "4.0"
deprecated==1.2.14 ; python_version >= "3.12" and python_version < "4.0"
google-api-core==2.19.2 ; python_version >= "3.12" and python_version < "4.0"
google-api-core[grpc]==2.19.2 ; python_version >= "3.12" and python_version < "4.0"
google-auth==2.34.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-appengine-logging==1.4.5 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-audit-log==0.3.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-bigquery-storage==2.26.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-bigquery==3.25.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-core==2.4.1 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-logging==3.11.2 ; python_version >= "3.12" and python_version < "4.0"
google-crc32c==1.6.0 ; python_version >= "3.12" and python_version < "4.0"
google-resumable-media==2.7.2 ; python_version >= "3.12" and python_version < "4.0"
googleapis-common-protos==1.65.0 ; python_version >= "3.12" and python_version < "4.0"
googleapis-common-protos[grpc]==1.65.0 ; python_version >= "3.12" and python_version < "4.0"
grpc-google-iam-v1==0.13.1 ; python_version >= "3.12" and python_version < "4.0"
grpcio-status==1.66.1 ; python_version >= "3.12" and python_version < "4.0"
grpcio==1.66.1 ; python_version >= "3.12" and python_version < "4.0"
idna==3.10 ; python_version >= "3.12" and python_version < "4.0"
importlib-metadata==8.4.0 ; python_version >= "3.12" and python_version < "4.0"
msgpack==1.0.8 ; python_version >= "3.12" and python_version < "4.0"
opentelemetry-api==1.27.0 ; python_version >= "3.12" and python_version < "4.0"
packaging==24.1 ; python_version >= "3.12" and python_version < "4.0"
proto-plus==1.24.0 ; python_version >= "3.12" and python_version < "4.0"
protobuf==5.28.1 ; python_version >= "3.12" and python_version < "4.0"
pyasn1-modules==0.4.1 ; python_version >= "3.12" and python_version < "4.0"
pyasn1==0.6.1 ; python_version >= "3.12" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.12" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.12" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.12" and python_version < "4"
six==1.16.0 ; python_version >= "3.12" and python_version < "4.0"
urllib3==2.2.3 ; python_version >= "3.12" and python_version < "4.0"
wrapt==1.16.0 ; python_version >= "3.12" and python_version < "4.0"
zipp==3.20.2 ; python_version >= "3.12" and python_version < "4.0"
